✨ Sometimes fetch a random kitten from Flickr
Since the curated list of kittens is pretty small, frequent visitors are
likely to run out of fresh kittens.

To prevent this, there’s now a random chance that a kitten will be
fetched from the Flickr API instead of the curated list.

Flickr’s search API generally returns pretty great kitten photos, but
every now and then something will sneak in that’s not actually a kitten.
Not much we can do about that, but you can just reload if that happens.