Improve license attribution for Flickr photos
We now display license attribution in the form "{title} by {author} is
licensed under {license name}". The title, author, and license name are
all links to the relevant URLs.

This is the attribution format recommended for images on the CC wiki:
<https://wiki.creativecommons.org/wiki/best_practices_for_attribution#This_is_a_great_attribution>

Thanks to Stephen Coles for pointing out that linking to the source
isn't sufficient: <https://mstdn.party/@stewf/109496153951715124>